ft(commands): add command crud
- add command controller and CRUD methods
- add command endpoints
- refactor categories controller delete method
- add helper functions

+++ server/controllers/categories.js
@@ -1,10 +1,15 @@
 import Category from '../models/Category';
-import { errorResponse } from '../helpers/utils';
+import Command from '../models/Command';
+import { errorResponse, notFoundError } from '../helpers/utils';
 
 const categoriesController = {
   create(req, res) {
     const { name } = req.sanitizedBody;
     return Category.findOne({ name }, (err, foundCategory) => {
+      if (err) {
+        return errorResponse(res);
+      }
+
       if (foundCategory) {
         return errorResponse(res,
           { statusCode: 422, message: 'Save failed. Category already exists.' });
@@ -27,7 +32,7 @@ const categoriesController = {
     });
   },
   fetchAll(req, res) {
-    return Category.find({}, (err, categories) => {
+    return Category.find({}, '_id, name', (err, categories) => {
       if (err) {
         return errorResponse(res);
       }
@@ -44,13 +49,9 @@ const categoriesController = {
   fetchOne(req, res) {
     const { id } = req.params;
 
-    return Category.findById(id, (err, category) => {
-      if (err) {
-        return errorResponse(res);
-      }
-
-      if (!category) {
-        return errorResponse(res, { statusCode: 404, message: 'Resource not found' });
+    return Category.findById(id, '_id, name', (err, category) => {
+      if (err || !category) {
+        return errorResponse(res, notFoundError);
       }
 
       return res.status(200).send({
@@ -67,12 +68,8 @@ const categoriesController = {
     const { name } = req.sanitizedBody;
 
     return Category.findById(id, (err, category) => {
-      if (err) {
-        return errorResponse(res);
-      }
-
-      if (!category) {
-        return errorResponse(res, { statusCode: 404, message: 'Resource not found' });
+      if (err || !category) {
+        return errorResponse(res, notFoundError);
       }
 
       return Category.findOne({ name }, (err, foundCategory) => {
@@ -99,16 +96,33 @@ const categoriesController = {
     });
   },
   delete(req, res) {
-    return Category.deleteOne({
-      _id: req.params.id
-    }, (err) => {
-      if (err) {
-        return errorResponse(res);
+    const { id } = req.params;
+    Category.findById(id, (err, category) => {
+      if (err || !category) {
+        return errorResponse(res, notFoundError);
       }
 
-      return res.status(200).send({
-        success: true,
-        message: 'Category deleted successfully'
+      Command.find({ category: category._id }, (err, commands) => {
+        if (err) {
+          return errorResponse(res);
+        }
+
+        if (commands.length !== 0) {
+          return errorResponse(res,
+            { message: 'Category contains children and cannot be deleted', statusCode: 422 });
+        }
+        return Category.deleteOne({
+          _id: id
+        }, (err) => {
+          if (err) {
+            return errorResponse(res);
+          }
+
+          return res.status(200).send({
+            success: true,
+            message: 'Category deleted successfully'
+          });
+        });
       });
     });
   }
